+++ pyproject.toml
@@ -22,3 +22,8 @@ jupyterlab = "^4.0.5"
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+pythonpath = [
+  "src"
+]
